Add initial version of Spack Site Generator package
spack_site_generator is a Python package for automating the creation of
site-specific YAML configuration files for Spack. It streamlines the process
of defining compilers, packages, modules, and global settings to ensure
consistency across environments. The generated YAMLs serve as a good starting
point for creating a Spack site on systems with exceptionally complex HPC
environments and system-specific configuration settings.