Add support for extra_attributes in package externals
Extended add_package to accept an extra_attributes dictionary and include
it in generated package externals. This enables passing paths such as
headers and libs when Spack recipes expect them but the package recipe
does not provide them (e.g., OpenMPI v5 with fftw). Tests were updated
and a new example was added to illustrate when extra_attributes must be
used.